Split diff section body rendering
